Avances en Update Articulo y arreglos varios
Se arreglan las funciones listar de Marca y Tipo.  Cambios en la consulta a la hora de hacer el Update

+++ J3AMS/J3AMS.Negocio/MarcaNegocio.cs
@@ -28,7 +28,7 @@ namespace J3AMS.Negocio
                 {
                     listMarcas.Add(new Marca()
                     {
-                        Id = (int)_datos.Lector["Id"],
+                        Id = (byte)_datos.Lector["Id"],
                         Descripcion = _datos.Lector["Descripcion"] as string ?? string.Empty,
                     });
                 }
